app/requirements.txt
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
gevent==21.12.0
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.7.1
parsel==1.6.0
Pillow==9.0.1
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyTelegramBotAPI==4.4.0
python-dotenv==0.19.2
queuelib==1.6.2
requests==2.27.1
schedule==1.1.0
Scrapy==2.5.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
Twisted==21.7.0
typing_extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
zope.event==4.5.0
zope.interface==5.4.0